+++ test/fixtures/clean/ok.xs
@@ -1,6 +1,9 @@
 // Clean example
 function "example" {
   input {
+    int page?=1
+    decimal weight?=1
+    int quantity?="-1"
   }
 
   stack {